WIP: Establish keyboard-navigation hooks
